refactor: lint and dry out code
