+++ src/lib/jobDescription.css
@@ -1,0 +1,8 @@
+.button {
+	flex: auto;
+	max-width: 10rem;
+	height: 2rem;
+	border-radius: 1rem;
+	padding-top: 0;
+	padding-bottom: 0;
+}
