+++ src/routes/Description.svelte
@@ -1,5 +1,7 @@
-<script>
+<script lang="ts">
 import { type Job } from '$lib/dbJobsTypes.js';
+import { listToArray, formatTime } from '$lib/jobDisplayFunctions.js';
+import '$lib/jobDescription.css';
 
 export let selected: Job;
 </script>
@@ -16,7 +18,7 @@ export let selected: Job;
 			</h3>
 			<p style="margin-bottom: 0">{selected.job}</p>
 		</header>
-	<ul>
+		<ul>
 			{#each listToArray(selected.short) as descLine}
 				<li><strong>{descLine.replace(/- /, '')}</strong></li>
 			{/each}
@@ -24,7 +26,7 @@ export let selected: Job;
 		<p>{selected.description}</p>
 		<p>
 			<em>{selected.location}</em>
-		<br />
+			<br />
 			<em
 				>{selected.days}: from {formatTime(selected.from)} to {formatTime(
 					selected.to,
@@ -42,5 +44,9 @@ dialog {
 }
 dialog > article {
 	padding-bottom: 2rem;
+}
+.close {
+	border: none;
+	box-shadow: none;
 }
 </style>
